Move collapseCrossGroupReferences into subfolder
